% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept-reject-submission.R
\name{accept_submission}
\alias{accept_submission}
\title{Accept a submission}
\usage{
accept_submission(syn, form_data_id)
}
\arguments{
\item{syn}{Synapse login object}

\item{form_data_id}{The formDataId of the submission.}
}
\value{
The response from the acceptance call as a list containing
  the submission metadata.
}
\description{
Accept a submission. This sets the submission status
`state` to `ACCEPTED`, and updates the review elements,
`reviewedOn` and `reviewedBy`. Note: this function will
throw an error if the REST call to Synapse fails; this could
happen if the user does not have needed permissions, or
the `form_data_id` specified does not exist.
}
\details{
WARNING: Once a submission is rejected, the state can
not be changed again.
}
