% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-update-delete-forms.R
\name{add_new_form}
\alias{add_new_form}
\title{Add a new form}
\usage{
add_new_form(syn, group, form_name, file_handle_id)
}
\arguments{
\item{syn}{Synapse login object}

\item{group}{The groupID.}

\item{form_name}{The name for the form.}

\item{file_handle_id}{A Synapse fileHandleId with the form
data.}
}
\description{
Add a new form to a group. User must have SUBMIT
permission on the ACL for the group.
See \href{https://docs.synapse.org/rest/POST/form/data.html}{Synapse REST API}. # nolint
}
