% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_json_to_table.R
\name{convert_all_submissions_to_table}
\alias{convert_all_submissions_to_table}
\title{Convert all json files in a directory to a single table.}
\usage{
convert_all_submissions_to_table(output_dir)
}
\arguments{
\item{output_dir}{The directory with the JSON files.}
}
\value{
A table with the column 'variables' that corresponds
  to all variable names across all JSON files, and each file's
  values are listed in a column of the table under the filename.
}
\description{
Convert all json files in a directory to a single table.
}
