% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-acl.R
\name{create_new_form_group}
\alias{create_new_form_group}
\title{Create new Synapse FormGroup}
\usage{
create_new_form_group(syn, name)
}
\arguments{
\item{syn}{Synapse login object}

\item{name}{Group name, between 3 and 256 characters,
that is globally unique. If a group with the same name exists
and the user has ACCESS_TYPE.READ permissions on the group,
then will get the FormGroup back.}
}
\value{
The FormGroup list, including the fields groupId,
  name, createdBy, and createdOn.
}
\description{
Create a new Synapse FormGroup. Creator will be
listed in ACL as administrator. See
\href{https://docs.synapse.org/rest/POST/form/group.html}{Synapse REST API}. # nolint
}
