% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-update-delete-forms.R
\name{delete_form}
\alias{delete_form}
\title{Delete a form}
\usage{
delete_form(syn, form_data_id)
}
\arguments{
\item{syn}{Synapse login object}

\item{form_data_id}{The formDataId for the form.}
}
\value{
None
}
\description{
Delete a form from the database, and caller must
have SUBMIT privileges in formGroup ACL.
}
