% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_all_and_get_table}
\alias{download_all_and_get_table}
\title{Get all submissions as single table}
\usage{
download_all_and_get_table(
  syn,
  group,
  state_filter = "SUBMITTED_WAITING_FOR_REVIEW"
)
}
\arguments{
\item{syn}{Synapse login object}

\item{group}{The groupID.}

\item{state_filter}{The filter that is desired to gather submissions by.}
}
\description{
Download all submissions, process into a single table. The column
names will be generated by [convert_all_forms_to_table()], which
currently is the formDataId given that it is unique.
}
