% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_all_and_output_to_csv}
\alias{download_all_and_output_to_csv}
\title{Get submissions and save to file}
\usage{
download_all_and_output_to_csv(
  syn,
  group,
  output_dir,
  state_filter = "SUBMITTED_WAITING_FOR_REVIEW"
)
}
\arguments{
\item{syn}{Synapse login object}

\item{group}{The groupID.}

\item{output_dir}{The directory to output the submissions and csv file to.}

\item{state_filter}{The filter that is desired to gather submissions by.}
}
\description{
Download all submissions, process into a single table, and output
the table to a csv file. This will also output all forms to the
same output directory.
}
