% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_submissions.R
\name{download_all_submissions}
\alias{download_all_submissions}
\title{Download all submission files to folder.}
\usage{
download_all_submissions(state_filter = "SUBMITTED_WAITING_FOR_REVIEW",
  group, output_dir)
}
\arguments{
\item{state_filter}{The filter that is desired to gather submissions by.}

\item{group}{The groupID.}

\item{output_dir}{The directory to output the files to.}
}
\description{
Download all submission files to folder.
}
