% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_submissions.R
\name{download_all_submissions_temp}
\alias{download_all_submissions_temp}
\title{Download all submissions}
\usage{
download_all_submissions_temp(
  syn,
  state_filter = "SUBMITTED_WAITING_FOR_REVIEW",
  group
)
}
\arguments{
\item{syn}{Synapse login object}

\item{state_filter}{The filter that is desired to gather submissions by.}

\item{group}{The groupID.}
}
\value{
List of temporary filenames.
}
\description{
Download all submissions to temporary files.
}
