% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms-to-synapse.R
\name{draft_message}
\alias{draft_message}
\title{Draft the body of an email message notifying of a form event}
\usage{
draft_message(exportable_forms, form_group_name, form_event)
}
\arguments{
\item{exportable_forms}{a dataframe as returned from `get_exportable_forms`.}

\item{form_group_name}{The name of the form group where the form event
took place.}

\item{form_event}{The form event. Possible values are
"create", "submit", or "review".}
}
\description{
Draft the body of an email message notifying of a form event
}
