% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms-to-synapse.R
\name{email_alert}
\alias{email_alert}
\title{Email a Synapse user/group in response to a form event}
\usage{
email_alert(
  syn,
  recipients,
  form_group_id,
  form_group_name,
  form_event = "submit",
  time_duration = 60,
  file_view_reference = NULL,
  as_reviewer = TRUE,
  submission_state = NULL
)
}
\arguments{
\item{syn}{A Synapse object (see `log_into_synapse`).}

\item{recipients}{A vector of recipients (Synapse user IDs) to send
an email alert to.}

\item{form_group_id}{The form group to check for the form event.}

\item{form_group_name}{The form group name. Currently (June 2020) the form
group ID is not sufficient information to fetch the form group and get
the group name so we need to manually pass the group name.}

\item{form_event}{The form event to check for. Possible values are
"create", "submit", or "review". "create" will consider when the form was added
to the form group whereas "submit" will consider when the form was submitted.}

\item{time_duration}{The time period (as an integer in seconds or a string
representation parseable by lubridate::duration) to consider a
form event "recent" and hence requiring an email alert. This serves
two purposes. First, if there is no file view specified (we are not exporting
form data to Synapse), this allows us to determine which forms events are recent
and which are not. Secondly, if there is a file view specified but exporting
the form data fails for some reason, this prevents the caller from repeatedly
emailing a user/group past a certain time window (when used with a cron job).
Set to Inf to fetch all qualifying events.}

\item{file_view_reference}{Synapse file view to query to check if form has
already been exported. If the form has already been exported there is
no email alert sent out.}

\item{as_reviewer}{Request forms using the /form/data/list/reviewer endpoint.
If FALSE, request forms using the /form/data/list endpoint. See the Synapse
REST docs for additional information.}
}
\description{
Email a Synapse user/group in response to a form event
}
