% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms-to-synapse.R
\name{get_exportable_forms}
\alias{get_exportable_forms}
\title{Get forms that don't yet exist in a file view on Synapse}
\usage{
get_exportable_forms(
  syn,
  form_group_id,
  file_view_reference,
  submission_state = "ACCEPTED",
  time_duration = Inf,
  form_event = "submit",
  as_reviewer = TRUE
)
}
\arguments{
\item{syn}{A Synapse object (see `log_into_synapse`).}

\item{form_group_id}{The form group to check for the form event.}

\item{file_view_reference}{Synapse file view to query to check if form has
already been exported. If NULL, this function behaves likes `get_recent_forms`}

\item{time_duration}{The time period (as an integer in seconds or a string
representation parseable by lubridate::duration) to consider a
form event "recent". By default it is Inf -- all forms are fetched.}

\item{form_event}{The form event to check for. Possible values are
"create", "submit", or "review". "create" will consider when the form was added
to the form group whereas "submit" will consider when the form was submitted.
Has no effect if time_duration is Inf.}

\item{as_reviewer}{Request forms using the /form/data/list/reviewer endpoint.
If False, request forms using the /form/data/list endpoint. See the Synapse
REST docs for additional information.}
}
\value{
A dataframe where each record is an exportable form.
}
\description{
Get forms that don't yet exist in a file view on Synapse
}
