% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-acl.R
\name{get_group_acl}
\alias{get_group_acl}
\title{Get FormGroup ACL}
\usage{
get_group_acl(syn, group)
}
\arguments{
\item{syn}{Synapse login object}

\item{group}{The FormGroup id number.}
}
\value{
The Access Control List (ACL) as a list.
}
\description{
Get the Access Control List for a FormGroup.
See \href{https://docs.synapse.org/rest/POST/form/group/id/acl.html}{Synapse REST API}. # nolint
}
