% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-submission-metadata.R
\name{get_json_as_df}
\alias{get_json_as_df}
\title{Get submission metadata as dataframe}
\usage{
get_json_as_df(data)
}
\arguments{
\item{data}{The json submission metadata.}
}
\description{
Get the submission metadata as a dataframe.
}
\keyword{internal}
