% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_json_to_table.R
\name{get_json_unlisted_table}
\alias{get_json_unlisted_table}
\title{Open a JSON file and return as a tibble.}
\usage{
get_json_unlisted_table(output_dir, filename)
}
\arguments{
\item{output_dir}{The directory the JSON file is in.}

\item{filename}{The name of the file.}
}
\value{
A tibble with the columns: variables, filename, where
  the variables are concatenated names from the JSON form,
  and the values for each variable are in the column filename.
}
\description{
Open a JSON file and return as a tibble.
}
