% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sub-info.R
\name{get_main_sections_tidier}
\alias{get_main_sections_tidier}
\title{Get main sections from tidier data}
\usage{
get_main_sections_tidier(data, submission_name)
}
\arguments{
\item{data}{Dataframe with at least two columns,
`variables` and `submission_name` if untidy. If `is_tidier`
is `TRUE`, then there should be columns `submission`, `section`,
and `response`.}

\item{submission_name}{Name of the submission.}
}
\description{
Get the high level sections from the form for a given
submission. Only returns sections that are not all NA.
}
\keyword{internal}
