% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms-to-synapse.R
\name{get_recent_forms}
\alias{get_recent_forms}
\title{Check for recent form events and return those forms}
\usage{
get_recent_forms(
  syn,
  form_group_id,
  time_duration,
  form_event = "create",
  as_reviewer = TRUE,
  submission_state = NULL
)
}
\arguments{
\item{syn}{A Synapse object (see `log_into_synapse`).}

\item{form_group_id}{The form group to check for the form event.}

\item{time_duration}{The time period (as an integer in seconds or a string
representation parseable by lubridate::duration) to consider a
form event "recent".}

\item{form_event}{The form event to check for. Possible values are
"create", "submit", or "review". "create" will consider when the form was added
to the form group, "submit" will consider when the form was submitted, and
"review" will consider when the form was reviewed. Has no effect if
time_duration is Inf.}

\item{as_reviewer}{Request forms using the /form/data/list/reviewer endpoint.
If False, request forms using the /form/data/list endpoint. See the Synapse
REST docs for additional information.}

\item{submission_state}{The submission state of the submissions.
Pass a list of submission states to match on any state in the list.
Submission states are: `WAITING_FOR_SUBMISSION`, `SUBMITTED_WAITING_FOR_REVIEW`,
`ACCEPTED`, `REJECTED`. Set to NULL (default) to ignore submission state.}
}
\value{
A dataframe where each record is a recent form.
}
\description{
Checks for either recent "create" events (when a form has been added to a
form group) or "submit" events (when a form has been submitted) and returns
forms which had the given event type occurr within a certain time_duration of
the current time. Due to how the Forms API behaves, we can only filter on the
created on time for forms that have already been submitted if the caller is
not the owner (creator) of a form.
}
