% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-json-to-table.R
\name{json_to_tidy}
\alias{json_to_tidy}
\title{Deserialize a JSON file to a tidy table}
\usage{
json_to_tidy(filepath, lists_are_atomic = TRUE)
}
\arguments{
\item{filepath}{The path to the JSON file}

\item{lists_are_atomic}{If TRUE, do not split lists elements into
separate fields. Instead, retain the list as a JSON elements.}
}
\description{
Convert a heirarchical JSON file to a tidy, tabular dataframe.
}
