% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_json_to_table.R
\name{main_section_column}
\alias{main_section_column}
\title{Get a table with main section as column}
\usage{
main_section_column(data)
}
\arguments{
\item{data}{The dataframe or tibble with the `variables`
column, where each variable is a string separated by
periods. The first string before the period is assumed
to indicate the main section.}
}
\value{
A dataframe or tibble where the `variables`
  have been broken into two columns, `main_section` and
  `variable`.
}
\description{
Get a table with the main section as a separate column.
This assumes the data is the unlisted data with variables
that are separated by periods. It takes the first string
before the period and puts it in a new column, `main_section`,
and puts the rest of the variable string in a `variable`
column.
}
