% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_submission_csv}
\alias{output_submission_csv}
\title{Write the data to a csv file}
\usage{
output_submission_csv(data, output_dir)
}
\arguments{
\item{data}{The data.}

\item{output_dir}{The directory to write the file to.}
}
\description{
Write the data to a csv file.
}
