% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept-reject-submission.R
\name{reject_submission}
\alias{reject_submission}
\title{Reject a submission}
\usage{
reject_submission(syn, form_data_id, reason)
}
\arguments{
\item{syn}{Synapse login object}

\item{form_data_id}{The formDataId of the submission.}

\item{reason}{The reason for the rejection as a string;
limited to 500 characters.}
}
\value{
The response from the rejection call as a list containing
  the submission metadata.
}
\description{
Reject a submission and give a reason. This sets the
submission status `state` to `REJECTED`, and updates the
review elements, `reviewedOn`, `reviewedBy`, and
`rejectionMessage`.` Note: this function will
throw an error if the REST call to Synapse fails; this could
happen if the user does not have needed permissions, or
the `form_data_id` specified does not exist.
}
\details{
WARNING: Once a submission is rejected, the state can
not be changed again.

See \href{https://docs.synapse.org/rest/POST/form/data/id/reject.html}{Synapse REST API}. # nolint
}
