% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest_delete}
\alias{rest_delete}
\title{restDELETE wrapper}
\usage{
rest_delete(syn, uri)
}
\arguments{
\item{syn}{Synapse login object}

\item{uri}{The uri for the API call}
}
\description{
Wrapper for the Synapse restDELETE API call.
}
