% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest_get}
\alias{rest_get}
\title{restGET wrapper}
\usage{
rest_get(syn, uri)
}
\arguments{
\item{syn}{Synapse login object}

\item{uri}{The uri for the API call}
}
\description{
Wrapper for the Synapse restGET API call.
}
