% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest_post}
\alias{rest_post}
\title{restPOST wrapper}
\usage{
rest_post(syn, uri, body = NULL)
}
\arguments{
\item{syn}{Synapse login object}

\item{uri}{The uri for the API call}

\item{body}{The body for the API call. Default is NULL since not all restPUT
API calls require a body.}
}
\description{
Wrapper for the Synapse restPOST API call.
}
