% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest_put}
\alias{rest_put}
\title{restPUT wrapper}
\usage{
rest_put(syn, uri, body)
}
\arguments{
\item{syn}{Synapse login object}

\item{uri}{The uri for the API call}

\item{body}{The body for the API call}
}
\description{
Wrapper for the Synapse restPUT API call.
}
