% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-update-delete-forms.R
\name{submit_form_for_review}
\alias{submit_form_for_review}
\title{Submit a form for review}
\usage{
submit_form_for_review(syn, form_data_id)
}
\arguments{
\item{syn}{Synapse login object}

\item{form_data_id}{The formDataId for the form.}
}
\description{
Submit a form for review. Warning: cannot update
form itself after it has been submitted. Only the
creator of the formDataId, if they have SUBMIT
permissions in the ACL can submit a form for review.
See \href{https://docs.synapse.org/rest/POST/form/data/id/submit.html}{Synapse REST API}. # nolint
}
