% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-group-acl.R
\name{update_group_acl}
\alias{update_group_acl}
\title{Update a FormGroup ACL}
\usage{
update_group_acl(syn, group, access_control_list)
}
\arguments{
\item{syn}{Synapse login object}

\item{group}{The FormGroup id number.}

\item{access_control_list}{The access control list (ACL) in JSON form.}
}
\description{
Update the ACL for a FormGroup.
See \href{https://docs.synapse.org/rest/PUT/form/group/id/acl.html}{Synapse REST API}. # nolint
WARNING: Ensure ACL is formatted correctly.
}
