% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms-to-synapse.R
\name{validate_form_event_params}
\alias{validate_form_event_params}
\title{Validate common parameters when filtering forms}
\usage{
validate_form_event_params(time_duration, form_event)
}
\arguments{
\item{time_duration}{The time period (as an integer in seconds or a string
representation parseable by lubridate::duration) to consider a
form event "recent".}

\item{form_event}{The recent event to check for. Possible values are
"create", "submit", or "review". "create" will trigger upon a new form being added
to the form group whereas "submit" will trigger upon an existing form being
submitted.}
}
\description{
Validate common parameters when filtering forms
}
