% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_all_and_output_to_csv}
\alias{download_all_and_output_to_csv}
\title{Download all submissions, process into a single table, and output
the table to a csv file.}
\usage{
download_all_and_output_to_csv(state_filter, group, output_dir)
}
\arguments{
\item{state_filter}{The filter that is desired to gather submissions by.}

\item{group}{The groupID.}

\item{output_dir}{The directory to output the submissions and csv file to.}
}
\description{
Download all submissions, process into a single table, and output
the table to a csv file.
}
