% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_submissions.R
\name{download_all_submissions_local}
\alias{download_all_submissions_local}
\title{Download all submissions locally}
\usage{
download_all_submissions_local(
  syn,
  state_filter = "SUBMITTED_WAITING_FOR_REVIEW",
  group,
  output_dir
)
}
\arguments{
\item{syn}{Synapse login object}

\item{state_filter}{The filter that is desired to gather submissions by.}

\item{group}{The groupID.}

\item{output_dir}{The directory to output the files to.}
}
\description{
Download all submissions to a local drive.
}
