% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-form-data.R
\name{download_form_file}
\alias{download_form_file}
\title{Download the submission file}
\usage{
download_form_file(ps_url, name, output_dir = ".")
}
\arguments{
\item{ps_url}{The presigned URL(s) for the submission.}

\item{name}{The basename of the submission file.}

\item{output_dir}{The directory to download the submission to.
By default the output directory is the working directory.}
}
\description{
Downloads the submission file locally.
}
