% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-json-to-table.R
\name{get_json_unlisted_table}
\alias{get_json_unlisted_table}
\title{Get data from JSON file}
\usage{
get_json_unlisted_table(filepath, name = NULL)
}
\arguments{
\item{filepath}{The path to the file, including name and extension.}

\item{name}{The desired column name for the submission data.
If `NULL`, the column will have a generic name.}
}
\value{
A tibble with the columns: variables, filename, where
  the variables are concatenated names from the JSON form,
  and the values for each variable are in the column filename.
}
\description{
Open a JSON file and return as a tibble with two columns,
the variables and the data. The data column name is `name` or
a generic name if not provided.
}
