% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sub-info.R
\name{get_main_sections}
\alias{get_main_sections}
\title{Get high level sections of form}
\usage{
get_main_sections(data, submission_name)
}
\arguments{
\item{data}{Dataframe with at least two columns,
`variables` and `submission_name`.}

\item{submission_name}{Name of the submission.}
}
\value{
A vector of the high level sections from
  the form for which the submission is not all NA.
  If `submission_name` is not in `names(data)`, then
  returns `NULL`.
}
\description{
Get the high level sections from the form for a given
submission. Only returns sections that are not all NA.
}
