% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-form-data.R
\name{get_ps_url}
\alias{get_ps_url}
\title{Get the presigned url for downloading the submission file.}
\usage{
get_ps_url(syn, file_handle_id, form_data_id)
}
\arguments{
\item{syn}{Synapse login object}

\item{file_handle_id}{The fileHandleId(s) for the submission(s).}

\item{form_data_id}{The formDataId(s) for the submission(s).}
}
\value{
The presigned URL(s).
}
\description{
Get the presigned url for downloading the submission file.
}
