% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sub-info.R
\name{get_submission_names}
\alias{get_submission_names}
\title{Get all submission names}
\usage{
get_submission_names(data, is_tidier = FALSE)
}
\arguments{
\item{data}{Dataframe with at least two columns,
`variables` and `submission_name` if untidy. If `is_tidier`
is `TRUE`, then there should be columns `submission`, `section`,
and `response`.}

\item{is_tidier}{True if `data` is in the form given by
[synapseforms::make_tidier_table()].}
}
\value{
A vector of the submission names.
}
\description{
Get all submission names.
}
