% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sub-info.R
\name{get_submission_names}
\alias{get_submission_names}
\title{Get all submission names}
\usage{
get_submission_names(data)
}
\arguments{
\item{data}{Dataframe or tibble with at least two columns,
`variables` and a submission with the name of the
submission as the column name.}
}
\value{
A vector of the submission names.
}
\description{
Get all submission names.
}
