% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submission_metadata.R
\name{get_submissions_metadata}
\alias{get_submissions_metadata}
\title{Get all submission metadata}
\usage{
get_submissions_metadata(
  syn,
  state_filter = "SUBMITTED_WAITING_FOR_REVIEW",
  group
)
}
\arguments{
\item{syn}{Synapse login object}

\item{state_filter}{The filter that is desired to gather submissions by.}

\item{group}{The groupID.}
}
\value{
A dataframe of the submission metadata.
}
\description{
Get the metadata for all submissions.
}
