% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_into_synapse}
\alias{log_into_synapse}
\title{Log into Synapse}
\usage{
log_into_synapse(username = NULL, password = NULL)
}
\arguments{
\item{username}{Your username}

\item{password}{Your password}
}
\value{
Synapse login object from
}
\description{
Log into Synapse.
}
