% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize-data.R
\name{make_tidier_table}
\alias{make_tidier_table}
\title{Make submissions table more tidy}
\usage{
make_tidier_table(data)
}
\arguments{
\item{data}{Data frame with the submissions.}
}
\description{
Make the submissions table more tidy by moving
column names to a key column and spreading out
the variables into three sections, "section",
"variable", and "sub_variable".
}
