% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-update-delete-forms.R
\name{update_form}
\alias{update_form}
\title{Update form}
\usage{
update_form(syn, form_data_id, file_handle_id, form_name = NULL)
}
\arguments{
\item{syn}{Synapse login object}

\item{form_data_id}{The formDataId for the form.}

\item{file_handle_id}{A Synapse fileHandleId with the form
data.}

\item{form_name}{The name for the form.}
}
\description{
Update a form. User must be the original owner
of the form and the state of the form must be either
`WAITING_FOR_SUBMISSION` or `REJECTED`. Froms with the states
`SUBMITTED_WAITING_FOR_REVIEW` or `ACCEPTED` are immutable.
See \href{https://docs.synapse.org/rest/PUT/form/data/id.html}{Synapse REST API}. # nolint
}
